Refactored api code 27/04
