+++ server/api/app/Http/Controllers/ApiCountryController.php
@@ -1,0 +1,14 @@
+
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+use App\Models\Country;
+
+
+class ApiCountryController extends Controller
+{
+
+}
